add schemas for score check test
