+++ app/api/v1/endpoints/score_check.py
@@ -1,9 +1,10 @@
 from pathlib import Path
 
-from fastapi import APIRouter, status
-from fastapi.responses import HTMLResponse, JSONResponse
-from pydantic import BaseModel
+from fastapi import APIRouter, HTTPException, status
+from fastapi.responses import HTMLResponse
 
+from app.schemas.score_check_input import ScoreCheckInput
+from app.schemas.score_check_response import ScoreCheckResponse, YakuScore
 from app.services.score_calculator.enums.enums import Tile, Wind
 from app.services.score_calculator.score_calculator import ScoreCalculator
 from tests.test_utils import (
@@ -15,18 +16,6 @@ from tests.test_utils import (
 router = APIRouter()
 
 
-class ScoreCheckInput(BaseModel):
-    raw_hand: str
-    winning_tile: str
-    is_discarded: bool
-    seat_wind: str = Wind.EAST.name
-    round_wind: str = Wind.EAST.name
-    is_last_tile_in_the_game: bool = False
-    is_last_tile_of_its_kind: bool = False
-    is_replacement_tile: bool = False
-    is_robbing_the_kong: bool = False
-
-
 @router.get("/score-check-test", response_class=HTMLResponse)
 def read_score_check_test() -> str:
     base_dir = Path(__file__).resolve().parent
@@ -34,27 +23,27 @@ def read_score_check_test() -> str:
     return html_path.read_text(encoding="utf-8")
 
 
-@router.post("/score-check")
-def score_check(input: ScoreCheckInput) -> JSONResponse:
+@router.post("/score-check", response_model=ScoreCheckResponse)
+def score_check(input: ScoreCheckInput) -> ScoreCheckResponse:
     hand = raw_string_to_hand_class(input.raw_hand)
     if (winning_tile := name_to_tile(input.winning_tile)) == Tile.F0:
-        return JSONResponse(
+        raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            content={"error": f"Invalid winning_tile: {input.winning_tile}"},
+            detail=f"Invalid winning_tile: {input.winning_tile}",
         )
     try:
         seat_wind = Wind[input.seat_wind]
     except KeyError:
-        return JSONResponse(
+        raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            content={"error": f"Invalid seat_wind: {input.seat_wind}"},
+            detail=f"Invalid seat_wind: {input.seat_wind}",
         )
     try:
         round_wind = Wind[input.round_wind]
     except KeyError:
-        return JSONResponse(
+        raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            content={"error": f"Invalid round_wind: {input.round_wind}"},
+            detail=f"Invalid round_wind: {input.round_wind}",
         )
 
     winning_conditions = create_default_winning_conditions(
@@ -70,13 +59,10 @@ def score_check(input: ScoreCheckInput) -> JSONResponse:
 
     score_calc = ScoreCalculator(hand=hand, winning_conditions=winning_conditions)
 
-    return JSONResponse(
-        status_code=status.HTTP_200_OK,
-        content={
-            "total_score": score_calc.result.total_score,
-            "yaku_score_list": [
-                {"name": yaku.name, "score": score}
-                for yaku, score in score_calc.result.yaku_score_list
-            ],
-        },
+    return ScoreCheckResponse(
+        total_score=score_calc.result.total_score,
+        yaku_score_list=[
+            YakuScore(name=yaku.name, score=score)
+            for yaku, score in score_calc.result.yaku_score_list
+        ],
     )
